pyproject.toml
[project]
name = "info-aggregator"
version = "0.1.0"
description = "个人信息聚合平台 MVP"
requires-python = ">=3.9.6"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4",
    "anthropic>=0.25",
    "jinja2>=3.1",
    "ghp-import>=2.1",
    "mistune>=3.0",
]

[project.scripts]
info-aggregator = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["collector", "storage", "analyzer", "publisher", "scheduler", "config"]
